Feat: Add program cloning
Adds ability to clone programs from mainnet when starting validators.
The `add-program` command allows users to add programs to the configuration file (`tp.config.json`).
When a validator is started, these programs will be cloned from mainnet and deployed to the localnet.

This feature enables testing against real-world programs, improving the accuracy and reliability of localnet testing.

The program cloning is enabled using validator's `--clone` and `--clone-upgradeable-program` flags.
It also prompts to reset existing validators when a program is added since the changes are applied during genesis.